60% lines
