+++ sergmiller/src/test/java/ru/fizteh/fivt/students/sergmiller/moduleTests/library/TwitterStreamTest.java
@@ -1,146 +1,102 @@
-//package ru.fizteh.fivt.students.sergmiller.moduleTests.library;
+package ru.mipt.diht.students.ale3otik.moduleTests.library;
 
 import com.beust.jcommander.JCommander;
-
-import org.hamcrest.core.IsCollectionContaining;
+import junit.framework.TestCase;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import junit.framework.TestCase;
-import org.junit.Test;
-import org.junit.Assert.*;
-
+import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
+import org.mockito.Mockito;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
-
-import ru.fizteh.fivt.students.sergmiller.twitterStream.*;
-import ru.fizteh.fivt.students.sergmiller.twitterStream.TwitterStream;
-import twitter4j.Status;
+import ru.fizteh.fivt.students.sergmiller.twitterStream.JCommanderParser;
+import ru.fizteh.fivt.students.sergmiller.twitterStream.LocationData;
+import ru.fizteh.fivt.students.sergmiller.twitterStream.TweetsGetterLimeted;
+import ru.fizteh.fivt.students.sergmiller.twitterStream.TwitterStreamLauncher;
+import sun.text.resources.cs.JavaTimeSupplementary_cs;
 import twitter4j.*;
-import twitter4j.Twitter4jTestUtils;
 
-import java.net.URL;
-import java.time.LocalDateTime;
-import java.time.Month;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Consumer;
 
 import static org.mockito.Mockito.*;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.hasItem;
-import static org.hamcrest.beans.HasPropertyWithValue.hasProperty;
-import static org.hamcrest.core.Is.is;
-import static org.hamcrest.core.IsCollectionContaining.hasItems;
-import static org.hamcrest.core.IsNot.not;
-import static org.junit.Assert.assertThat;
-import static org.mockito.AdditionalMatchers.or;
-import static org.mockito.Matchers.*;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 /**
- * Created by sergmiller on 03.11.15.
+ * Created by sergmiller on 15.11.15.
  */
 
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({TimeResolver.class})
-//@RunWith(MockitoJUnitRunner.class)
+@PrepareForTest({Thread.class})
 public class TwitterStreamTest extends TestCase{
     private LocationData LondonLocation = new LocationData(
             new GeoLocation(51.5073509, -0.1277583), 23.539304731202712, "London");
-    private static List<Status> doctorWhoTweets;
-
-
-    //@Mock
-    //private QueryResult queryResultDoctorWho;
-
-    @Mock
-    Twitter twitter;
-
-    @Mock
-    GeoLocationResolver geoLocationResolver;
+    private LocationData DolgoprudniyLocation = new LocationData(
+            new GeoLocation(55.947064, 37.4992755), 6.117792942260596, "Dolgoprudnyy");
+    private StatusAdapter statusAdapter;
+    private JCommanderParser jCommanderParser;
+    private JCommander jCommander;
+    private TwitterStreamLauncher twitterStreamLauncher;
+    List <String> results = new ArrayList<>();
+    public static List <Status> doctorWhoTweets;
+    private Consumer<String> dummyConsumer = (x)->mockedPrint(x);
 
     @Mock
-    twitter4j.TwitterStream twitterStream;
+    private Status dummyStatus;
 
-//    @Mock
-//    LocalDateTime localDateTime;
+    @Mock
+    private TwitterStream dummyTwitterStream;
 
     @BeforeClass
     public static void loadStatuses() {
         doctorWhoTweets = Twitter4jTestUtils.tweetsFromJson("/DoctorWhoInLondonTweets.json");
     }
 
-    @Before
-    public void setUp() throws Exception {
-        when(geoLocationResolver.getGeoLocation("London"))
-               .thenReturn(LondonLocation);
+//    @Before
+//    public void setUp() {
+//        ArgumentCaptor<StatusListener> statusCaptor
+//                = ArgumentCaptor.forClass(StatusListener.class);
+//        doNothing().when(twitterStream).addListener((StatusListener)
+//                statusCaptor.capture());
+//        doAnswer(invocation -> {
+//            aMoscowStatuses.forEach(s -> statusCaptor.getValue().onStatus(s));
+//            return null;
+//        }).when(twitterStream).filter(any(FilterQuery.class));
+//        Mockito.when(dummyStatus.getGeoLocation()).thenReturn(null);
+//        Mockito.when(dummyStatus.getText()).thenReturn("some message");
 //
-        QueryResult queryResultDoctorWho = mock(QueryResult.class);
-        when(queryResultDoctorWho.getTweets()).thenReturn(doctorWhoTweets);
+//        PowerMockito.mockStatic(Thread.class);
+//
+//        twitterStreamLauncher = new TwitterStreamLauncher(dummyTwitterStream, dummyConsumer);
+//
+//
+//    }
 
-        when(twitter.search(argThat(hasProperty("query", equalTo("doctorWho")))))
-                .thenReturn(queryResultDoctorWho);
-
-        PowerMockito.mockStatic(TimeResolver.class);
-        PowerMockito.when(TimeResolver.getTime(any(LocalDateTime.class), any(LocalDateTime.class))).thenReturn("Только что");
-    }
-
-    @Test
-    public void testPrintTwitterLimitedAllTweets() throws Exception {
-        final String[] dummyArgs = {"-q", "doctorWho", "-p", "London"};
-        JCommanderParser doctorWhoQuery = new JCommanderParser();
-        JCommander jCommander = new JCommander(doctorWhoQuery, dummyArgs);
-        TwitterStream twitterStream7 = new TwitterStream();
-        List<String> currentTweets = twitterStream7.getTwitterLimited(
-                        doctorWhoQuery, LondonLocation, twitter);
-
-        assertEquals(5, currentTweets.size());
-
-//        for (int i = 0;i < 5;++i) {
-//            System.out.println(currentTweets.get(i));
+//    private void createLauncherWithArguments(boolean isGeolocationNeeded,String... args) {
+//        arguments = new Arguments();
+//        jcm = new JCommander(arguments);
+//        jcm.parse(args);
+//        if(isGeolocationNeeded){
+//            arguments.setGeoLocationInfo(londonGeoLocationInfo);
 //        }
-        assertThat(currentTweets, hasItem("[Только что] @\u001B[34mJay_Cree\u001B[0m: #Doctorwho #thedoctor"
-                + " #london #excellondon #doctorwhofestival #drwhofestival #drwho"
-                + " #british #english… https://t.co/gWjxU63YTm (0 ретвитов)\n"
-                + "--------------------------------------------------------------------------------"));
+//
+//    }
 
-        assertThat(currentTweets, hasItem("[Только что] @\u001B[34mDoctorWhoVic\u001B[0m:"
-                + " ретвитнул @\u001B[34mAnnCarters\u001B[0m:"
-                + "  Early designs for the #robotofsherwood robots."
-                + " #dwfestuk #doctorwho #MFX @ ExCeL London https://t.co/YRJ5Dqlazj\n"
-                + "--------------------------------------------------------------------------------"));
+    public void mockedPrint(String string) {
+        results.add(string);
     }
 
-    @Test
-    public void testPrintTwitterLimitedTweetsWithoutRetweets() throws Exception {
-        final String[] dummyArgs = {"-q", "doctorWho", "-p", "London", "--hideRetweets"};
-        JCommanderParser doctorWhoQuery = new JCommanderParser();
-        JCommander jCommander = new JCommander(doctorWhoQuery, dummyArgs);
-        TwitterStream twitterStream7 = new TwitterStream();
-        List<String> currentTweets = twitterStream7.getTwitterLimited(
-                doctorWhoQuery, LondonLocation, twitter);
-
-        assertEquals(3, currentTweets.size());
-
-//        for (int i = 0;i < 3;++i) {
-//            System.out.println(currentTweets.get(i));
-//        }
-
-        assertThat(currentTweets, hasItem("[Только что] @\u001B[34malexryans\u001B[0m:"
-                + " Final day in London at the #DoctorWho Festival"
-                + " - looking forward to seeing all the cast and writers!… "
-                + "https://t.co/BExJ7mYvGd (0 ретвитов)\n"
-                + "--------------------------------------------------------------------------------"));
-
-        assertThat(currentTweets, not(hasItem("[Только что] @\u001B[34mDoctorWhoVic\u001B[0m:"
-                + " ретвитнул @\u001B[34mAnnCarters\u001B[0m:"
-                + "  Early designs for the #robotofsherwood robots."
-                + " #dwfestuk #doctorwho #MFX @ ExCeL London https://t.co/YRJ5Dqlazj\n"
-                + "--------------------------------------------------------------------------------")));
-    }
+//    @Test
+//    public void testPrintTwitterStream() throws Exception{
+//        final String[] dummyArgs = {"-q", "doctorWho", "-p", "London", "-s"};
+//        JCommanderParser doctorWhoQuery = new JCommanderParser();
+//        JCommander jCommander = new JCommander(doctorWhoQuery, dummyArgs);
+//        twitterStreamLauncher.printTwitterStream(doctorWhoQuery, LondonLocation);
+//
+//        //assertThat();
+//    }
 
 }
